backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
httpx
Pillow
passlib[bcrypt]
python-jose[cryptography]
ollama
tensorflow